# vps/opt-timetrack/timetrack/timetrack-web-prototype/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Timetrack.kz API
  description: |
    REST API для сервиса учёта рабочего времени Timetrack.kz.

    ## Аутентификация
    Все защищённые эндпоинты требуют передачи заголовка `Authorization: Bearer <token>`.
    Токен получается через `POST /api/auth/login`.

    ## Часовой пояс
    Все datetime-поля в ISO 8601 с явным offset, например: `2026-06-16T09:02:00+05:00`

    ## Версионирование
    Текущая версия: v1. Базовый URL: `https://api.timetrack.kz/v1`
  version: "1.0.0"
  contact:
    name: Timetrack.kz Support
    url: https://timetrack.kz
    email: support@timetrack.kz
servers:
  - url: https://api.timetrack.kz/v1
    description: Production
  - url: https://staging.api.timetrack.kz/v1
    description: Staging
  - url: http://localhost:3000/api
    description: Local development (Next.js mock)
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Вход в систему
      description: Возвращает access token и refresh token. Используется как веб-панелью, так и мобильным приложением.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@timetrack.kz
              password: password
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Обновить access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Новый access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      summary: Выход из системы
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Успешный выход
  /auth/me:
    get:
      tags: [auth]
      summary: Получить текущего пользователя
      description: Используется мобильным приложением для проверки токена и получения профиля.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'

  # COMPANY ───────────────────────────────────
...
  /employees:
    get:
      tags: [employees]
      summary: Список сотрудников
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 25 }
        - name: search
          in: query
          schema: { type: string }
          description: Поиск по имени, email, коду
        - name: departmentId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: workScheduleId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated list of employees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeListResponse'

    post:
      tags: [employees]
      summary: Создать сотрудника
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreateRequest'
      responses:
        '201':
          description: Созданный сотрудник
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
  /employees/import:
    post:
      tags: [employees]
      summary: Импорт сотрудников из Excel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Результат импорта
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  skipped:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: string
  /employees/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [employees]
      summary: Получить сотрудника
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Данные сотрудника
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [employees]
      summary: Обновить сотрудника
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdateRequest'
      responses:
        '200':
          description: Обновлённый сотрудник
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'

    delete:
      tags: [employees]
      summary: Удалить / деактивировать сотрудника
      description: Реализуется как soft delete — status меняется на inactive
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Успешно деактивирован

  # LOCATIONS ─────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Неверный или истёкший токен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Seconds until expiry
        user:
          $ref: '#/components/schemas/AuthUser'
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, COMPANY_ADMIN, HR, MANAGER, EMPLOYEE]
        companyId:
          type: string
        employeeId:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true

    # ── COMPANY ─────────────────────
...
    Employee:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        employeeCode:
          type: string
        departmentId:
          type: string
        position:
          type: string
        managerId:
          type: string
          nullable: true
        workScheduleId:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        hiredAt:
          type: string
          format: date
        avatarUrl:
          type: string
          nullable: true
        hasFaceTemplate:
          type: boolean
        createdAt:
          type: string
          format: date-time
    EmployeeCreateRequest:
      type: object
      required: [fullName, email, departmentId, position, employeeCode]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        employeeCode:
          type: string
        departmentId:
          type: string
        position:
          type: string
        managerId:
          type: string
        workScheduleId:
          type: string
        hiredAt:
          type: string
          format: date
    EmployeeUpdateRequest:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        departmentId:
          type: string
        position:
          type: string
        workScheduleId:
          type: string
        status:
          type: string
          enum: [active, inactive]
        managerId:
          type: string
    EmployeeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    # ── LOCATION ─────────────────────
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
security:
  - bearerAuth: []